Move exception in exceptions file
