+++ nilearn/_utils/__init__.py
@@ -1,6 +1,6 @@
 
 from .niimg_conversions import (check_niimg, check_niimg_3d, concat_niimgs,
-    check_niimg_4d, DimensionError)
+    check_niimg_4d)
 
 from .niimg import new_img_like, _repr_niimgs, copy_img
 
@@ -9,3 +9,5 @@ from .numpy_conversions import as_ndarray
 from .cache_mixin import CacheMixin
 
 from .logger import _compose_err_msg
+
+from .exceptions import DimensionError
